Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,9 +5,9 @@ char* welcome() {
 }
 
 int main() {
-    int a, b,c;
-    scanf("%d %d %d",&a,&b,&c);
-    int average = (a + b + c) / 3;
+    float a, b,c;
+    scanf("%f %f %f",&a,&b,&c);
+    float average = (a + b + c) / 3;
     printf("Average: %.2f",average);
     return 0;
 }
